+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/CoinJoinProgressEventArgs.cs
@@ -1,0 +1,5 @@
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class CoinJoinProgressEventArgs : EventArgs
+{
+}
