+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/EnteringInputRegistrationPhase.cs
@@ -1,0 +1,10 @@
+using WalletWasabi.WabiSabi.Models;
+
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class EnteringInputRegistrationPhase : RoundStateChanged
+{
+	public EnteringInputRegistrationPhase(RoundState roundState, DateTimeOffset timeout) : base(roundState, timeout)
+	{
+	}
+}
