+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/EnteringSigningPhase.cs
@@ -1,0 +1,10 @@
+using WalletWasabi.WabiSabi.Models;
+
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class EnteringSigningPhase : RoundStateChanged
+{
+	public EnteringSigningPhase(RoundState roundState, DateTimeOffset timeout) : base(roundState, timeout)
+	{
+	}
+}
