+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/RoundEnded.cs
@@ -1,0 +1,13 @@
+using WalletWasabi.WabiSabi.Models;
+
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class RoundEnded : CoinJoinProgressEventArgs
+{
+	public RoundEnded(RoundState roundState)
+	{
+		RoundState = roundState;
+	}
+
+	public RoundState RoundState { get; }
+}
