+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/WaitingForBlameRound.cs
@@ -1,0 +1,11 @@
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class WaitingForBlameRound : CoinJoinProgressEventArgs
+{
+	public WaitingForBlameRound(DateTimeOffset dateTimeOffset)
+	{
+		DateTimeOffset = dateTimeOffset;
+	}
+
+	public DateTimeOffset DateTimeOffset { get; }
+}
