+++ WalletWasabi/WabiSabi/Client/CoinJoinProgressEvents/WaitingForRound.cs
@@ -1,0 +1,5 @@
+namespace WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
+
+public class WaitingForRound : CoinJoinProgressEventArgs
+{
+}
