+++ WalletWasabi/WabiSabi/Client/Events/CoinJoinStatusEventArgs.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using WalletWasabi.WabiSabi.Client.CoinJoinProgressEvents;
 using WalletWasabi.Wallets;
 
 namespace WalletWasabi.WabiSabi.Client.Events;
